# docs/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Denguekay: pronóstico distrital de dengue",
    "description": "API del documento OE2 (Tabla 2) sobre datos agregados de Piura. Las predicciones de versiones que no cumplen los umbrales se marcan experimental: true (demo; desviación temporal documentada).",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/admin/inferencias": {
      "post": {
        "tags": [
          "Administración"
        ],
        "summary": "Ejecutar la inferencia de una semana de corte (HU0009)",
        "description": "Reconstruye los vectores desde la BD, aplica la versión en uso y hace UPSERT de predicciones y alertas.\n\nNo entrena. h=3, un corte sin observaciones o un artefacto alterado responden 409.",
        "operationId": "inferencias_api_v1_admin_inferencias_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudInferencia"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenciaRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "Clave de escritura ausente o inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Solicitud no atendible con los datos o modelos (h=3, artefacto, umbrales)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Parámetro inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "503": {
            "description": "Base de datos o escritura no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Error interno controlado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/modelos/{version}/activar": {
      "post": {
        "tags": [
          "Administración"
        ],
        "summary": "Activar una versión que cumple los umbrales (HU0009)",
        "description": "Archiva la versión activa de la misma tarea y horizonte, en una transacción auditada.\n\nUna versión con cumple_umbrales = false no se activa (409): así lo exige el DDL.",
        "operationId": "activar_api_v1_admin_modelos__version__activar_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Identificador positivo del registro guardado en la BD",
              "title": "Version"
            },
            "description": "Identificador positivo del registro guardado en la BD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivacionRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "Clave de escritura ausente o inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Solicitud no atendible con los datos o modelos (h=3, artefacto, umbrales)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Parámetro inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "503": {
            "description": "Base de datos o escritura no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Error interno controlado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivacionRespuesta": {
        "properties": {
          "id_ejecucion": {
            "type": "integer",
            "title": "Id Ejecucion"
          },
          "activada": {
            "$ref": "#/components/schemas/ModeloRespuesta"
          },
          "archivada": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModeloRespuesta"
              },
              {
                "type": "null"
              }
            ]
          },
          "nota": {
            "type": "string",
            "title": "Nota"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id_ejecucion",
          "activada",
          "archivada",
          "nota"
        ],
        "title": "ActivacionRespuesta",
        "description": "Activación auditada como ejecución de mantenimiento"
      },
...
      "ErrorRespuesta": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "detalle": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalle"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "codigo",
          "mensaje"
        ],
        "title": "ErrorRespuesta",
        "description": "Error homogéneo, sin credenciales ni SQL interno."
      },
...
      "InferenciaRespuesta": {
        "properties": {
          "id_ejecucion": {
            "type": "integer",
            "title": "Id Ejecucion"
          },
          "horizonte": {
            "type": "integer",
            "title": "Horizonte"
          },
          "id_semana_corte": {
            "type": "integer",
            "title": "Id Semana Corte"
          },
          "id_semana_objetivo": {
            "type": "integer",
            "title": "Id Semana Objetivo"
          },
          "experimental": {
            "type": "boolean",
            "title": "Experimental"
          },
          "seleccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seleccion"
          },
          "versiones": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Versiones"
          },
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          },
          "predicciones": {
            "type": "integer",
            "title": "Predicciones"
          },
          "disponibles": {
            "type": "integer",
            "title": "Disponibles"
          },
          "no_disponibles": {
            "type": "integer",
            "title": "No Disponibles"
          },
          "alertas": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Alertas"
          },
          "duracion_segundos": {
            "type": "number",
            "title": "Duracion Segundos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id_ejecucion",
          "horizonte",
          "id_semana_corte",
          "id_semana_objetivo",
          "experimental",
          "seleccion",
          "versiones",
          "motivo",
          "predicciones",
          "disponibles",
          "no_disponibles",
          "alertas",
          "duracion_segundos"
        ],
        "title": "InferenciaRespuesta",
        "description": "Resumen de la ejecución de inferencia confirmada en la BD"
      },
...
      "ModeloRespuesta": {
        "properties": {
          "id_version": {
            "type": "integer",
            "title": "Id Version"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "tarea": {
            "type": "string",
            "enum": [
              "clasificacion",
              "regresion"
            ],
            "title": "Tarea"
          },
          "horizonte": {
            "type": "integer",
            "title": "Horizonte"
          },
          "algoritmo": {
            "type": "string",
            "title": "Algoritmo"
          },
          "estado": {
            "type": "string",
            "enum": [
              "candidata",
              "activa",
              "archivada",
              "rechazada"
            ],
            "title": "Estado"
          },
          "cumple_umbrales": {
            "type": "boolean",
            "title": "Cumple Umbrales"
          },
          "experimental": {
            "type": "boolean",
            "title": "Experimental"
          },
          "umbral_probabilidad": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Umbral Probabilidad"
          },
          "mlflow_run_id": {
            "type": "string",
            "title": "Mlflow Run Id"
          },
          "fecha_registro": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Registro"
          },
          "fecha_activacion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Activacion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id_version",
          "codigo",
          "tarea",
          "horizonte",
          "algoritmo",
          "estado",
          "cumple_umbrales",
          "experimental",
          "umbral_probabilidad",
          "mlflow_run_id",
          "fecha_registro",
          "fecha_activacion"
        ],
        "title": "ModeloRespuesta"
      },
...
      "SolicitudInferencia": {
        "properties": {
          "horizonte": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 2.0,
            "title": "Horizonte",
            "description": "2 o 4; 3 responde 409 (sin modelo)"
          },
          "id_semana_corte": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 209953.0,
                "minimum": 201701.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Semana Corte",
            "description": "Semana de corte (anio*100+semana); por defecto, la última observada"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "horizonte"
        ],
        "title": "SolicitudInferencia"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Clave de escritura definida en API_KEY; las lecturas son públicas",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}